requirements.txt
djangorestframework==3.12.4
Django==3.2
pytz==2021.1
django-extensions==3.1.3
sqlparse==0.4.1
psycopg2